Develop in the Kubernetes cluster via SKAFFY SKAFFY SKAFFOLD
I'm in between laptops, so reached the awkward stage of deciding if it's best to
install mysql locally (e.g. homebrew) or via docker-compose. Fortunately I
remembered Skaffold lets me develop in the Kubernetes cluster as-if I'm
developing on my local machine.

Therefore I don't need to provision databases as the in-dev container can talk
straight to the hosted database. Develop against a prod databases y'all.